Complete Stage 3
Add a menu to the window of the text editor. The Load and Save items
should do the same as the buttons of the same name. The Exit item
should quit the program. Add fields for the load/save buttons so that
we can access them from the menu item methods.